+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:name="com.cjwsc.idcm.base.application.BaseApplication"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".MainActivity">
+        <activity android:name=".MainKTActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -18,7 +18,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="com.foxexchange.user.MyAssetActivity"
+        <activity android:name="com.foxexchange.user.MyAssetKTActivity"
             android:screenOrientation="locked"/>
     </application>
