+++ app/src/main/java/com/foxexchange/android/fragments/ExchangeKTFragment.kt
@@ -1,11 +1,9 @@
 package com.foxexchange.android.fragments
 
 import android.os.Bundle
-import com.cjwsc.idcm.base.BaseView
-import com.cjwsc.idcm.kotlin.base.BaseFragment
+import com.cjwsc.idcm.kotlin.base.BaseKTFragment
 
-class MarketFragment : BaseFragment() {
-
+class ExchangeKTFragment : BaseKTFragment() {
     override val layoutId: Int
         get() = 0
