--- common/src/main/java/com/foxexchange/common/activitys/BaseFoxExChangeActivity.kt
@@ -1,5 +1,0 @@
-package com.foxexchange.common.activitys
-
-import com.cjwsc.idcm.base.BaseActivity
-
-abstract class BaseFoxExChangeActivity : com.cjwsc.idcm.kotlin.base.BaseActivity()
