+++ common/src/main/java/com/foxexchange/common/activitys/BaseFoxExChangeKTActivity.kt
@@ -1,0 +1,3 @@
+package com.foxexchange.common.activitys
+
+abstract class BaseFoxExChangeKTActivity : com.cjwsc.idcm.kotlin.base.BaseKTActivity()
