+++ common/src/main/java/com/foxexchange/common/fragments/BaseFoxExChangeFragment.kt
@@ -1,9 +1,6 @@
 package com.foxexchange.common.fragments
 
-import android.os.Bundle
-import com.cjwsc.idcm.kotlin.base.BaseFragment
-
-// abstract class BaseFoxExChangeFragment : BaseFragment()
+// abstract class BaseFoxExChangeFragment : BaseKTFragment()
 // {
 //
 //     override val layoutId: Int
