--- user/src/main/Java/com.foxexchange.user/fragments/AssetListFragment.kt
@@ -1,27 +1,0 @@
-package com.foxexchange.user.fragments
-
-import android.os.Bundle
-import com.cjwsc.idcm.base.BaseFragment
-import com.example.user.R
-
-/**
- * 资产列表fragment
- */
-class AssetListFragment  : com.cjwsc.idcm.kotlin.base.BaseFragment() {
-
-    override val layoutId: Int
-        get() = R.layout.fragment_asset_list_layout
-
-    override fun onInitView(bundle: Bundle?) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun onEvent() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun lazyFetchData() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-}
