+++ user/src/main/Java/com.foxexchange.user/fragments/AssetListKTFragment.kt
@@ -1,0 +1,24 @@
+package com.foxexchange.user.fragments
+
+import android.os.Bundle
+import com.cjwsc.idcm.kotlin.base.BaseKTFragment
+import com.example.user.R
+
+/**
+ * 资产列表fragment
+ */
+class AssetListKTFragment  : BaseKTFragment() {
+
+    override val layoutId: Int
+        get() = R.layout.fragment_asset_list_layout
+
+    override fun onInitView(bundle: Bundle?) {
+    }
+
+    override fun onEvent() {
+    }
+
+    override fun lazyFetchData() {
+    }
+
+}
