--- user/src/main/Java/com.foxexchange.user/fragments/C2CAssetFragment.kt
@@ -1,26 +1,0 @@
-package com.foxexchange.user.fragments
-
-import android.os.Bundle
-import com.cjwsc.idcm.kotlin.base.BaseFragment
-import com.example.user.R
-
-/**
- * c2c资产
- */
-class C2CAssetFragment : BaseFragment() {
-
-    override val layoutId: Int
-        get() = R.layout.fragment_c2c_asset_layout
-
-    override fun onInitView(bundle: Bundle?) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun onEvent() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun lazyFetchData() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-}
