+++ user/src/main/Java/com.foxexchange.user/fragments/C2CAssetKTFragment.kt
@@ -1,0 +1,23 @@
+package com.foxexchange.user.fragments
+
+import android.os.Bundle
+import com.cjwsc.idcm.kotlin.base.BaseKTFragment
+import com.example.user.R
+
+/**
+ * c2c资产
+ */
+class C2CAssetKTFragment : BaseKTFragment() {
+
+    override val layoutId: Int
+        get() = R.layout.fragment_c2c_asset_layout
+
+    override fun onInitView(bundle: Bundle?) {
+    }
+
+    override fun onEvent() {
+    }
+
+    override fun lazyFetchData() {
+    }
+}
