--- user/src/main/Java/com.foxexchange.user/fragments/CoinAssetFragment.kt
@@ -1,25 +1,0 @@
-package com.foxexchange.user.fragments
-
-import android.os.Bundle
-import com.cjwsc.idcm.kotlin.base.BaseFragment
-import com.example.user.R
-
-/**
- * 币币资产fragment
- */
-class CoinAssetFragment : BaseFragment() {
-    override val layoutId: Int
-        get() = R.layout.fragment_coin_asset_layout
-
-    override fun onInitView(bundle: Bundle?) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun onEvent() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun lazyFetchData() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-}
