+++ user/src/main/Java/com.foxexchange.user/fragments/CoinAssetKTFragment.kt
@@ -1,0 +1,22 @@
+package com.foxexchange.user.fragments
+
+import android.os.Bundle
+import com.cjwsc.idcm.kotlin.base.BaseKTFragment
+import com.example.user.R
+
+/**
+ * 币币资产fragment
+ */
+class CoinAssetKTFragment : BaseKTFragment() {
+    override val layoutId: Int
+        get() = R.layout.fragment_coin_asset_layout
+
+    override fun onInitView(bundle: Bundle?) {
+    }
+
+    override fun onEvent() {
+    }
+
+    override fun lazyFetchData() {
+    }
+}
